Rename namerd interpreter classes out of io.l5d package
