+++ interpreter/namerd/src/main/resources/META-INF/services/io.buoyant.namer.InterpreterInitializer
@@ -1,1 +1,1 @@
-io.l5d.NamerdInterpreterInitializer
+io.buoyant.namerd.iface.NamerdInterpreterInitializer
